fix offense of Metrics/ClassLength: by creating two spearate class
