+++ author.json
@@ -1,1 +1,1 @@
-[{"first_name":"Stephen","last_name":"King"}]
+[{"first_name":"Stephen","last_name":"King"},{"first_name":"Leo","last_name":"Tolstoy"}]
