+++ book.json
@@ -1,1 +1,1 @@
-[{"Publisher":"Macmillan Publishers","cover_state":"Good","publisher_date":"2023-02-02","genre":"Comedy","label_title":"Gift","label_color":"New","author_firstname":"Stephen","author_lasttname":"King"}]
+[{"Publisher":"Macmillan Publishers","cover_state":"Good","publisher_date":"2023-02-02","genre":"Comedy","label_title":"Gift","label_color":"New","author_firstname":"Stephen","author_lasttname":"King"},{"Publisher":"John Wiley and Sons","cover_state":"Good","publisher_date":"2022-05-10","genre":"Thriller","label_title":"Suny","label_color":"New","author_firstname":"Leo","author_lasttname":"Tolstoy"}]
