+++ genre.json
@@ -1,1 +1,1 @@
-[{"name":"Comedy"}]
+[{"name":"Comedy"},{"name":"Thriller"}]
