+++ label.json
@@ -1,1 +1,1 @@
-[{"title":"Gift","color":"New"}]
+[{"title":"Gift","color":"New"},{"title":"Suny","color":"New"}]
